+++ include/TUIC/point2.h
@@ -105,7 +105,7 @@ int tuiPoint2GetCrossProductZ(const TuiPoint2 point2_1, const TuiPoint2 point2_2
  *
  * @thread_safety This function does not access memory in ways that could cause a data race, so it is thread safe.
  */
-float tuiPoint2GetDotProduct(const TuiPoint2 point2_1, const TuiPoint2 point2_2);
+int tuiPoint2GetDotProduct(const TuiPoint2 point2_1, const TuiPoint2 point2_2);
 /*! @} */
